Pull emitter configuration into SparkleImage
